Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    double a;
-    scanf("%lf",&a);
+    float a;
+    scanf("%f",&a);
     printf("Your entered: %.2f",a);
     return 0;
